More INIT and PN stuff
The app now tells the server when a user follows/unfollows a class
(defined in global.js). it also tells the server when a new user should
be recorded in the DB. Added info to localStorage to determine if it is
someone's first time launching the app. Also added the username and
registration ID to localStorage.